Conflicts:
	snakemake/version.py

+++ snakemake/script.py
@@ -151,7 +151,7 @@ class Snakemake:
 
 
 def script(path, basedir, input, output, params, wildcards, threads, resources,
-           log, config, rulename, conda_env):
+           log, config, rulename, conda_env, bench_record):
     """
     Load a script from the given basedir + path and execute it.
     Supports Python 3 and R.
@@ -170,6 +170,7 @@ def script(path, basedir, input, output, params, wildcards, threads, resources,
     else:
         sourceurl = path
 
+    f = None
     try:
         with urlopen(sourceurl) as source:
             if path.endswith(".py"):
@@ -184,7 +185,7 @@ def script(path, basedir, input, output, params, wildcards, threads, resources,
                 import sys; sys.path.insert(0, "{}"); import pickle; snakemake = pickle.loads({})
                 ######## Original script #########
                 """).format(searchpath, snakemake)
-            elif path.endswith(".R"):
+            elif path.endswith(".R") or path.endswith(".Rmd"):
                 preamble = textwrap.dedent("""
                 ######## Snakemake header ########
                 library(methods)
@@ -226,7 +227,7 @@ def script(path, basedir, input, output, params, wildcards, threads, resources,
                            }), REncoder.encode_dict(config), REncoder.encode_value(rulename))
             else:
                 raise ValueError(
-                    "Unsupported script: Expecting either Python (.py) or R (.R) script.")
+                    "Unsupported script: Expecting either Python (.py), R (.R) or RMarkdown (.Rmd) script.")
 
             if path.startswith("file://"):
                 # in case of local path, use the same directory
@@ -242,8 +243,22 @@ def script(path, basedir, input, output, params, wildcards, threads, resources,
                 prefix=prefix,
                 dir=dir,
                 delete=False) as f:
-                f.write(preamble.encode())
-                f.write(source.read())
+                if not path.endswith(".Rmd"):
+                    f.write(preamble.encode())
+                    f.write(source.read())
+                else:
+                    # Insert Snakemake object after the RMarkdown header
+                    code = source.read().decode()
+                    pos = code.rfind("---")
+                    f.write(str.encode(code[:pos+3]))
+                    preamble = textwrap.dedent("""
+                        ```{r, echo=FALSE, message=FALSE, warning=FALSE}
+                        %s
+                        ```
+                        """ % preamble)
+                    f.write(preamble.encode())
+                    f.write(str.encode(code[pos+3:]))
+
             if path.endswith(".py"):
                 py_exec = sys.executable
                 if conda_env is not None:
@@ -263,10 +278,18 @@ def script(path, basedir, input, output, params, wildcards, threads, resources,
                                         "master process to execute "
                                         "script.".format(*MIN_PY_VERSION))
                 # use the same Python as the running process or the one from the environment
-                shell("{py_exec} {f.name}")
+                shell("{py_exec} {f.name}", bench_record=bench_record)
             elif path.endswith(".R"):
-                shell("Rscript {f.name}")
-            os.remove(f.name)
+                shell("Rscript {f.name}", bench_record=bench_record)
+            elif path.endswith(".Rmd"):
+                if len(output) != 1:
+                    raise WorkflowError("RMarkdown scripts (.Rmd) may only have a single output file.")
+                out = os.path.abspath(output[0])
+                shell("Rscript -e 'rmarkdown::render(\"{f.name}\", output_file=\"{out}\", quiet=TRUE, params = list(rmd=\"{f.name}\"))'",
+                    bench_record=bench_record)
 
     except URLError as e:
         raise WorkflowError(e)
+    finally:
+        if f:
+            os.remove(f.name)
